Option for joint to exceed limit as warn
Previously, whenever a joint exceeded its upper or lower limit, it produced an error, if the amount was larger than specified in the limit_tolerance rosparam. Now, exceeding any limit by a joint will produce a warning, unless the limit is exceeded by an amount larger than the limit_tolerance, in which case an error is produced.